adding authentication

+++ pages/api/seed.ts
@@ -1,6 +1,7 @@
 import type { NextApiRequest, NextApiResponse } from "next";
 import { db, seedDatabase } from "../../database";
 import { Product } from "../../models";
+import User from "../../models/user";
 
 type Data = {
   message: string;
@@ -10,11 +11,14 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse<Data>
 ) {
-  // if (process.env.NODE_ENV === "production") {
-  //   return res.status(401).json({ message: "No tiene acceso a este API" });
-  // }
-
+  if (process.env.NODE_ENV === "production") {
+    return res.status(401).json({ message: "No tiene acceso a este API" });
+  }
   await db.connect();
+
+  await User.deleteMany();
+  await User.insertMany(seedDatabase.initialData.users);
+
   await Product.deleteMany();
   await Product.insertMany(seedDatabase.initialData.products);
